refactor: extract default filter fns
