+++ src/app/(core)/teachers/programmes/[slug]/[tab]/page.tsx
@@ -30,7 +30,6 @@ import withPageErrorHandling, {
 } from "@/hocs/withPageErrorHandling";
 import { resolveOakHref } from "@/common-lib/urls";
 import { getSubjectPhaseSlug } from "@/components/TeacherComponents/helpers/getSubjectPhaseSlug";
-import { getDefaultFilter } from "@/utils/curriculum/filtering";
 import { BrowseFiltersProvider } from "@/context/BrowseFilters";
 import { redirectProgrammeSlugIfNeeded } from "@/utils/integratedJourney/legacyProgrammeUnitsRedirect";
 import { cacheData } from "@/node-lib/cache";
@@ -39,7 +38,7 @@ import { getMvRefreshTime } from "@/pages-helpers/curriculum/downloads/getMvRefr
 import { validateServerSearchParams } from "@/utils/validateProgrammePageSearchParams";
 import { TeacherBrowseAnalyticsStoreProvider } from "@/context/TeacherBrowseAnalytics/TeacherBrowseAnalyticsProvider";
 import { getProgrammeStateForProgramme } from "@/context/TeacherBrowseAnalytics/utils/getProgrammeState";
-import { resolveBrowseFilterFromSearchParams } from "@/context/BrowseFilters/utils";
+import { getBrowseFilterState } from "@/context/BrowseFilters/utils/getBrowseFilterState";
 
 const reportError = errorReporter("programme-page::app");
 
@@ -252,11 +251,10 @@ const InnerProgrammePage = async (props: AppPageProps<ProgrammePageParams>) => {
     formatCurriculumUnitsData(curriculumUnitsData);
 
   // Resolve filter server-side from URL search params
-  const defaultFilter = getDefaultFilter(curriculumUnitsFormattedData);
-  const resolvedFilter = resolveBrowseFilterFromSearchParams(
-    curriculumUnitsFormattedData,
+  const { defaultFilter, resolvedFilter } = getBrowseFilterState({
+    data: curriculumUnitsFormattedData,
     searchParams,
-  );
+  });
 
   // All KS4 options for subject phase
   const ks4Options =
