+++ src/context/BrowseFilters/utils/getBrowseFilterState.ts
@@ -1,0 +1,23 @@
+import { getDefaultBrowseFilter } from "./getDefaultBrowseFilter";
+import {
+  RawSearchParams,
+  resolveBrowseFilterFromSearchParams,
+} from "./resolveBrowseFilterFromSearchParams";
+
+import { CurriculumUnitsFormattedData } from "@/pages-helpers/curriculum/docx/tab-helpers";
+
+export const getBrowseFilterState = ({
+  data,
+  searchParams,
+}: {
+  data: CurriculumUnitsFormattedData;
+  searchParams?: RawSearchParams;
+}) => {
+  const defaultFilter = getDefaultBrowseFilter(data);
+  const resolvedFilter = resolveBrowseFilterFromSearchParams(
+    data,
+    searchParams,
+  );
+
+  return { defaultFilter, resolvedFilter };
+};
