+++ src/utils/curriculum/filtering.test.ts
@@ -3,7 +3,6 @@ import {
   getDefaultSubjectCategoriesForYearGroup,
   getDefaultTiersForYearGroup,
   isHighlightedUnit,
-  getDefaultFilter,
   getFilterData,
   scopeYearsToKeystageFilter,
   shouldDisplayFilter,
@@ -250,43 +249,6 @@ describe("filtering", () => {
     expect(isHighlightedUnit(unit, [thread4.slug])).toBeFalsy();
     expect(isHighlightedUnit(unit, [])).toBeFalsy();
     expect(isHighlightedUnit(unit, null)).toBeFalsy();
-  });
-
-  describe("getDefaultFilter", () => {
-    it("with data", () => {
-      const out = getDefaultFilter({
-        yearData: {
-          "7": {
-            units: [] as Unit[],
-            tiers: [{ tier_slug: "foundation", tier: "Foundation" }],
-            childSubjects: [{ subject: "Physics", subject_slug: "physics" }],
-            subjectCategories: [
-              createSubjectCategory({ id: 2, slug: "sub-cat-2" }),
-            ],
-          } as CurriculumUnitsYearData[number],
-          "8": {
-            units: [] as Unit[],
-            tiers: [{ tier_slug: "higher", tier: "Higher" }],
-            childSubjects: [{ subject: "Biology", subject_slug: "biology" }],
-            subjectCategories: [
-              createSubjectCategory({ id: 1, slug: "sub-cat-1" }),
-            ],
-          } as CurriculumUnitsYearData[number],
-        },
-        threadOptions: [],
-        yearOptions: ["7", "8"],
-        keystages: [],
-      });
-      expect(out).toEqual({
-        childSubjects: ["biology"],
-        subjectCategories: ["sub-cat-1"],
-        threads: [],
-        tiers: ["foundation"],
-        years: ["7", "8"],
-        pathways: [],
-        keystages: [],
-      });
-    });
   });
 });
